+++ resources/views/forms/components/layup-builder.blade.php
@@ -15,9 +15,6 @@
                 widgetRegistry: @js($widgetRegistry),
                 translations: @js($translations),
                 initialLivePreviews: @js($initialLivePreviews),
-                geocode: function() {
-                console.log('tf');
-                }
             })"
 
             {{ $getExtraAttributeBag() }}